# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is an application web of interview assessment tests for interviewing out of the box .",
        "title": "test",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/my-tests": {
            "post": {
                "description": "create new Test by json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "add new Test",
                "parameters": [
                    {
                        "description": "Add Test",
                        "name": "Test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Test"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TestResponse"
                        }
                    }
                }
            }
        },
...
        "/my-tests/clone/{id}": {
            "post": {
                "description": "clone test",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "clone a test",
                "operationId": "cloneTest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "test id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "expected Time",
                        "name": "expectedTime",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CloneTestInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MyTests"
                        }
                    }
                }
            }
        },
...
        "/my-tests/{id}": {
            "get": {
                "description": "get test by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "get test by id",
                "operationId": "getTest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "test id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MyTests"
                        }
                    }
                }
            }
        },
...
        "/my-tests/{test_id}": {
            "post": {
                "description": "update Test by json and path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "update Test",
                "operationId": "updateTest",
                "parameters": [
                    {
                        "description": "Update Test",
                        "name": "Test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TestRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Update Test",
                        "name": "test_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TestResponse"
                        }
                    }
                }
            }
        },
...
        "/tests": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "get tests by skill",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "get tests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "tests search by type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TestResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Answer": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "answer_choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnswerChoices"
                    }
                },
                "answer_file": {
                    "type": "string"
                },
                "answer_text": {
                    "type": "string"
                },
                "candidate_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "point": {
                    "type": "number"
                },
                "question_id": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.AnswerChoices": {
            "type": "object",
            "required": [
                "choices_id"
            ],
            "properties": {
                "answer_id": {
                    "type": "integer"
                },
                "choices_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Candidate": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "test": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Test"
                    }
                }
            }
        },
...
        "models.Choices": {
            "type": "object",
            "required": [
                "choice_text",
                "is_answer"
            ],
            "properties": {
                "choice_text": {
                    "type": "string"
                },
                "is_answer": {
                    "type": "boolean"
                }
            }
        },
        "models.CloneTestInput": {
            "type": "object",
            "properties": {
                "expected_time": {
                    "type": "integer"
                }
            }
        },
...
        "models.MyTests": {
            "type": "object",
            "properties": {
                "expected_time": {
                    "type": "integer"
                },
                "number_Question": {
                    "type": "integer"
                },
                "number_candidate": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "integer"
                },
                "test_name": {
                    "type": "string"
                }
            }
        },
        "models.Question": {
            "type": "object",
            "required": [
                "name",
                "question_text"
            ],
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Answer"
                    }
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Choices"
                    }
                },
                "difficulty": {
                    "type": "string"
                },
                "expected_time": {
                    "type": "integer"
                },
                "file_read_me": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "max_points": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "question_text": {
                    "type": "string"
                },
                "skill": {
                    "$ref": "#/definitions/models.Skill"
                },
                "skill_id": {
                    "type": "integer"
                },
                "test": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Test"
                    }
                },
                "test_questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TestQuestion"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Skill": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Test": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "candidate": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Candidate"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notify_emails": {
                    "type": "string"
                },
                "passing_score": {
                    "type": "integer"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Question"
                    }
                },
                "show_score": {
                    "type": "boolean"
                },
                "time_limit": {
                    "type": "integer"
                },
                "timing_policy": {
                    "type": "string"
                }
            }
        },
...
        "models.TestQuestion": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.TestRequest": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notify_emails": {
                    "type": "string"
                },
                "passing_score": {
                    "type": "integer"
                },
                "show_score": {
                    "type": "boolean"
                },
                "timing_policy": {
                    "type": "string"
                }
            }
        },
        "models.TestResponse": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notify_emails": {
                    "type": "string"
                },
                "passing_score": {
                    "type": "integer"
                },
                "show_score": {
                    "type": "boolean"
                },
                "timing_policy": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}